requirements.txt
fastapi
uvicorn
python-dotenv
mistralai
pydantic
Jinja2
aiofiles
python-multipart
mistral
reportlab